scheduler: flip around optout check
If the schedule.xml (incorrectly) doesn't contain a recording.output
field, it is going to be represented as the empty string, which fails
the 'eq "false"' test and causes all recordings to be considered optout.

Flip the check around so we record an event if it has no optout property
defined.