scheduler: automatically reload schedule
Originally, we required manual triggering of schedule reloads to protect
us from fuckups caused by haphazard schedule updates. However, this
turned out not to be as much of an issue, while forgetting to trigger
a reload is.

Now, we reload the schedule based on its mtime. This is safe because
get-fahrplan does not update the schedule in place, but creates a new
file and moves it over the old one.