genpage: drop talks which are not in the schedule
If the schedule changes and removes a talk which relive has already
recorded, genpage will generate an index.json entry without any
metadata, which doesn't look too nice for the clients.

As those talks are probably not supposed to show up anymore anyway, we
simply don't put them in the index anymore.